* Moved Service Tags into ServiceTags component. * Added Realm selector to Service Search page.
