+++ src/modules/route-services/components/index.ts
@@ -2,3 +2,4 @@ export * from './nav.component';
 export * from './search-result.component';
 export * from './service-listing.component';
 export * from './service-notification.component'
+export * from './service-tags.component'
