+++ src/modules/route-services/pages/search.page.tsx
@@ -1,25 +1,12 @@
 import { t } from '@lingui/macro';
 import { useLingui } from '@lingui/react';
-import { Box, Card, Divider, Grid, ToggleButton, ToggleButtonGroup, Typography } from '@mui/material';
+import { Box, Card, Divider, Grid, Typography } from '@mui/material';
 import React from 'react';
-import { SearchResult } from '../components';
-
-const TAGS = {
-    POWERLEVELING: 1 << 0,
-    BOSS_HELP: 1 << 1,
-    UBER_LILITH: 1 << 2,
-    CAPSTONE_BOOST: 1 << 3,
-    EUROPE: 1 << 4,
-    ASIA: 1 << 5,
-    AMERICA: 1 << 6,
-};
+import { SearchResult, ServiceTags } from '../components';
+import { ServerTypeInput } from '@modules/common/components';
+import { useRouteServerType } from '../../common/providers';
 
 export const SearchPage: React.FC = () => {
-    const [selectedTags, setSelectedTags] = React.useState<number[]>([]);
-
-    const handleTagSelection = (_, newTags: number[]) => {
-        setSelectedTags(() => newTags);
-    };
 
     // const tagsToNumber = React.useMemo(
     //     () =>
@@ -29,6 +16,8 @@ export const SearchPage: React.FC = () => {
     //         ),
     //     [selectedTags]
     // );
+
+    const [serverType, setServerType] = useRouteServerType();
 
     const { i18n } = useLingui();
 
@@ -43,20 +32,14 @@ export const SearchPage: React.FC = () => {
                             </Typography>
                             <Divider />
                         </Grid>
-                        <Grid item xs={12} lg={6}>
-                            <ToggleButtonGroup
-                                value={selectedTags}
-                                onChange={handleTagSelection}
-                                aria-label={t(i18n)`Service Types`}
-                            >
-                                <ToggleButton value={TAGS.POWERLEVELING}>{t(i18n)`Powerleveling`}</ToggleButton>
-                                <ToggleButton value={TAGS.BOSS_HELP}>{t(i18n)`Boss Help`}</ToggleButton>
-                                <ToggleButton value={TAGS.UBER_LILITH}>{t(i18n)`Uber Lilith`}</ToggleButton>
-                                <ToggleButton value={TAGS.CAPSTONE_BOOST}>{t(i18n)`Capstone Boost`}</ToggleButton>
-                                <ToggleButton value={TAGS.EUROPE}>{t(i18n)`Europe`}</ToggleButton>
-                                <ToggleButton value={TAGS.ASIA}>{t(i18n)`Asia`}</ToggleButton>
-                                <ToggleButton value={TAGS.AMERICA}>{t(i18n)`America`}</ToggleButton>
-                            </ToggleButtonGroup>
+                        <Grid item xs={9}>
+                            <ServiceTags />
+                        </Grid>
+                        <Grid item xs={3}>
+                            <ServerTypeInput
+                                value={serverType}
+                                onChange={setServerType}                                            
+                            />                            
                         </Grid>
                     </Grid>
                 </Box>
